Move auth. logic from RootLayout to AuthorizationGate
